Added Dictionary Unpacking
Changed from manually entering the inflection features for verbs into creating nested dictionaries. This process is slower to run but can be extrapolated to more inflections in the future. However, it may be possible to do an iterrows() function, remove the NaN rows and then unpack the left over ones into fname